Moves mdotdevs into mdot completely.
