+++ mdot/templates/mdot/developers/guidelines.html
@@ -1,15 +1,15 @@
-{% extends "mdotdevs/base.html" %}
+{% extends "mdot/developers/base.html" %}
 
 {% block title %}Building Apps Home{% endblock %}
 
 {% block content %}
 
 <div class="mdot-developers pub-guidelines">
-    
+
     <div class="container">
-        
+
         <h2 class="mdot-subhead" id="page_header">UW Developer Store <span class="nobreak">Publishing Guidelines</span></h2>
-        
+
         <ol class="trad-list">
             <li class="article"><h3>Meet campus needs</h3>The app must meet known needs of broad campus-wide audiences. If it is intended for very specific audiences (such as a class, department, or school), we advise departments and schools to set up their own App Stores.  If you need assistance in doing this, please feel free to contact us at <a href="mailto:help@uw.edu">help@uw.edu</a>.</li>
             <li class="article"><h3>Sponsorship & support</h3>The app must have a clearly identified:
@@ -31,9 +31,9 @@
                     <li><a href="http://www.washington.edu/accessibility/help/">Get help</a>. We highly recommend scheduling a free consultation on your app with <a href="mailto:help@uw.edu">UW-IT Accessible Technology Services (ATS)</a>. </li>
                 </ul>
             </li>
-            <li class="article"><h3>UW brand alignment </h3>Using the UW logo? If so, review the <a href="http://www.washington.edu/accessibility/guidelines/">UW Brand Guidelines</a> for appropriate use guidelines.</li> 
+            <li class="article"><h3>UW brand alignment </h3>Using the UW logo? If so, review the <a href="http://www.washington.edu/accessibility/guidelines/">UW Brand Guidelines</a> for appropriate use guidelines.</li>
         </ol>
-        
+
     </div>
 </div>
