+++ mdot/templates/mdot/developers/home.html
@@ -1,19 +1,19 @@
-{% extends "mdotdevs/base.html" %}
+{% extends "mdot/developers/base.html" %}
 
 {% block title %}Building Apps Home{% endblock %}
 
 {% block content %}
 
 <div class="mdot-developers">
-    
+
     <div class="container" >
-        
+
         <h2 class="mdot-subhead" id="page_header">Building mobile apps</h2>
-        
+
         <h3>Things to consider</h3>
-        
-        <div class="row"> 
-                                  
+
+        <div class="row">
+
             <div class="col-md-4 article">
                 <h4>Mobile web vs. Native vs. Hybrid</h4><p>Before deciding what to build, understand your options. Mobile web is the most versatile and easiest to maintain, but can be limiting if you need robust device integration. </p>
                 <ul>
@@ -29,19 +29,19 @@
                     <li><a title="Forbes: Know When You Should Build Custom Software Over Canned Solutions" href="http://www.forbes.com/sites/chuckcohn/2014/09/15/build-vs-buy-how-to-know-when-you-should-build-custom-software-over-canned-solutions/">Build vs. Buy</a></li>
                     <li><a title="Boxes and Arrows: Understanding the Market Context" href="http://boxesandarrows.com/competitive-analysis-understanding-the-market-context/">Doing a Competitive Analysis</a></li>
                 </ul>
-            </div> 
+            </div>
             <div class="col-md-4 article">
                 <h4>True Cost</h4><p>Offering a software solution is never without cost. Understand the real costs for ongoing development, maintenance, support, and communication.</p>
                 <ul>
                     <li><a title="SFCD Blog: How much does it cost to make an app?" href="https://sfcd.com/blog/cost/">The Cost of an App</a></li>
                 </ul>
             </div>
-         
+
         </div>
-        
+
         <h3>Create a successful product</h3>
-        
-        <div class="row"> 
+
+        <div class="row">
             <div class="col-md-4 article">
                 <h4>Solve the Right Problems, for Real Users</h4><p>Understand and prioritize your user population(s), their needs, and their goals for your software. </p>
                 <ul>
@@ -65,14 +65,14 @@
                     <li><a title="Intercom Blog: Ways to Improve A Product" href="https://blog.intercom.io/ways-to-improve-a-product/">Ways to Improve A Product</a></li>
                 </ul>
             </div>
-         
+
         </div>
-        
+
 
 <h3>Development Concerns</h3>
-        
-        <div class="row"> 
-                                  
+
+        <div class="row">
+
             <div class="col-md-4 article">
                 <h4>So Many Options</h4><p>Choosing the technology upon which to build is a huge step. Understand your various options.</p>
                 <ul>
@@ -96,7 +96,7 @@
                     <li><a title="Ilya Grigorik: Optimizing the Critical Renderign Path for Instant Mobile Websites [video]" href="https://www.youtube.com/watch?v=YV1nKLWoARQ">Optimization for Mobile Websites [Video]  </a></li>
                 </ul>
             </div>
-         
+
         </div>
