+++ mdot/templates/mdot/developers/process.html
@@ -1,13 +1,13 @@
-{% extends "mdotdevs/base.html" %}
+{% extends "mdot/developers/base.html" %}
 
 {% block title %}Building Apps Home{% endblock %}
 
 {% block content %}
 
 <div class="mdot-developers">
-    
+
     <div class="container" >
-        
+
         <h2 class="mdot-subhead" id="page_header">UW Developer Store <span class="nobreak">Publishing Process</span></h2>
         <p class="article intro">So, you have developed a native app and want to publish it under the UW Developer Store? Here is what you do next. </p>
         <ol class="trad-list process-page">
@@ -26,9 +26,9 @@
                     <li>Publication, pending app store approval.</li>
                 </ul>
             </li>
-            
+
         </ol>
-        
+
     </div>
 </div>
