+++ mdot/templates/mdot/developers/thanks.html
@@ -1,18 +1,18 @@
-{% extends "mdotdevs/base.html" %}
+{% extends "mdot/developers/base.html" %}
 
 {% block title %}Building Apps Home{% endblock %}
 
 {% block content %}
 
 <div class="mdot-developers">
-    
+
     <div class="container" >
-        
+
         <h2 class="mdot-subhead" id="page_header">Thank you</h2>
-        <p class="thanks">Your app will be reviewed for inclusion in the UW Development App Store. The review process takes around 2 weeks.</p> 
+        <p class="thanks">Your app will be reviewed for inclusion in the UW Development App Store. The review process takes around 2 weeks.</p>
         <p class="thanks">We will contact the app sponsor about next steps.</p>
-        
-        
+
+
     </div>
 </div>
 {% endblock %}
