+++ mdot/tests.py
@@ -7,3 +7,4 @@ from mdot.test.client_error_catching import MdotClientErrorTest
 from mdot.test.client_get_resources import MdotClientRequestTest
 from mdot.test.client_helper_methods import MdotClientMethodsTest
 from mdot.test.url_and_view import MdotTest
+from mdot.test.forms import MdotFormTest
