+++ mdot/urls.py
@@ -8,6 +8,10 @@ urlpatterns = patterns(
     # Examples:
 
     url(r'^$', 'mdot.views.home', name='home'),
+    url(r'^developers/$', 'mdot.views.developers', name='developers'),
+    url(r'^developers/guidelines/$', 'mdot.views.guidelines', name='guidelines'),
+    url(r'^developers/process/$', 'mdot.views.process', name='process'),
+    url(r'^developers/review/$', 'mdot.views.review', name='review'),
 
     # url(r'^blog/', include('blog.urls')),
     # url(r'^admin/', include(admin.site.urls)),
