--- mdotdevs/tests.py
@@ -1,4 +1,0 @@
-from django.utils import unittest
-
-from mdotdevs.test.url_and_view import MdotdevTest
-from mdotdevs.test.forms import MdotdevsFormTest
