+++ travis-ci/urls.py
@@ -19,9 +19,8 @@ from django.conf import settings
 from django.conf.urls.static import static
 
 urlpatterns = [
-    url(r'^admin/', include(admin.site.urls)),
+    # url(r'^admin/', include(admin.site.urls)),
     # url(r'^api/v1/', include('mdot_rest.urls')),
-    url(r'^developers/', include('mdotdevs.urls')),
     url(r'^', include('mdot.urls')),
 ]
